fix seed

+++ repository/pgsql/courier_cost_mapping.go
@@ -8,5 +8,5 @@ type CourierCostMapping struct {
 	PostcodeCitySource      string   `gorm:"index:idx_member" json:"postcode_city_source"`
 	PostcodeCityDestination string   `gorm:"index:idx_member" json:"postcode_city_destination"`
 	BasePrice               int      `json:"base_price"`
-	CoefWeight              float64  `json:"coef_weight"`
+	CoefWeight              int      `json:"coef_weight"`
 }
